Replayed packets have no payload but always full flow authenticator

+++ elements/local/castor/castor.hh
@@ -78,10 +78,10 @@ public:
 	inline Hash* fauth() { return reinterpret_cast<Hash*>((uint8_t*) this + sizeof(*this) + ((syn) ? sizeof(Nonce) : 0)); }
 
 	inline unsigned int header_len() const {
-		return sizeof(*this) + (syn ? sizeof(Nonce) : 0) + fasize * hsize;
+		return sizeof(*this) + (syn ? sizeof(Nonce) : 0) + (unsigned int) fasize * hsize;
 	}
 	inline unsigned int payload_len() const {
-		return ntohs(len) - header_len();
+		return (unsigned int) ntohs(len) - header_len();
 	}
 };
